NEW - bug 376976: [Widgets] Various bugfixes/improvements https://bugs.eclipse.org/bugs/show_bug.cgi?id=376976
made by anaïs lefeuvre

+++ plugins/infra/widget/org.eclipse.papyrus.infra.widgets/src/org/eclipse/papyrus/infra/widgets/editors/DoubleEditor.java
@@ -58,9 +58,9 @@ public class DoubleEditor extends StringEditor {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public Float getValue() {
+	public Double getValue() {
 		try {
-			return (Float)targetToModelConverter.convert(super.getValue());
+			return (Double)targetToModelConverter.convert(super.getValue());
 		} catch (Exception ex) {
 			Activator.log.error(ex);
 			return null;
